automation/requirements.txt
# ResLibs 自动化脚本依赖
# 核心依赖
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3

# AI 内容生成
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6

# 百度网盘 API
bypy==1.9.9
requests-oauthlib==1.3.1

# 图片搜索和下载
unsplash-python==1.0.0
pexels-python==0.1.0
pixabay-sdk==1.0.2
Pillow==10.1.0

# Cloudflare R2 存储
boto3==1.34.0
botocore==1.34.0
awscli==2.13.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
prisma==0.10.0

# 文件处理
patoolib==2.0.0
py7zr==0.21.0
python-magic==0.4.27

# 文件上传
tuspy==1.1.0
uploadlib==0.1.0

# 日志和监控
structlog==23.2.0
colorlog==6.8.0
rich==13.7.0

# 数据处理
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# 网络和代理
urllib3==2.1.0
certifi==2023.11.17
python-socks==2.4.3

# 定时任务
schedule==1.2.0
apscheduler==3.10.4

# 通知
slack-sdk==3.26.1
discord-webhook==1.3.0
smtplib-ssl==1.0.0

# 实用工具
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.2

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 安全和加密
cryptography==41.0.8
pycryptodome==3.19.0